\name{output}
\docType{data}
\alias{output}
\title{triM output for variance component & FIA analyses of the example dataset.}
\description{
 triM output for variance component & FIA analyses of the example dataset, which is used in MCIBD, MCIBD.chro and MCIBD.genome functions.
 }
\format{ This dataset contains a triM output format object named output, as input of MCIBD, MCIBD.chro and MCIBD.genome functions.
}
\keyword{datasets}